OS name more flexible
the OS name used to be hardcoded to only work in distros that started the ``NAME=`` field in ``/etc/os-release`` with a quotation mark. since testing gentoo and other distributions, this does not work, as their fields do not have quotation marks for the ``NAME=`` field. This could be fixed using the ``PRETTY_NAME`` field but unlike the ``NAME`` field, it shows up in different spots throughout the file, and this would mean extensive searching through the file to find the right field.